+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 - Implemented `sll_arg_state_t`, `sll_free_args` and `sll_parse_args_list`
 - Implemented `sll_complex_abs`, `sll_complex_add`, `sll_complex_conjugate`, `sll_complex_div`, `sll_complex_div_float`, `sll_complex_mult`, `sll_complex_mult_float`, `sll_complex_neg`, `sll_complex_pow`, `sll_complex_pow_float`, `sll_complex_pow_int`, `sll_complex_reciprocal` and `sll_complex_sub` ([#262])
 - Implemented `sll_complex_t` ([#262])
-- Implemented `SLL_ERROR_FLAG_SLL` and `SLL_LIBC_UNMAPPED_WIN_ERROR`
+- Implemented `SLL_ERROR_FLAG_SLL`, `SLL_LIBC_UNMAPPED_WIN_ERROR`, `SLL_ERROR_EOF` and `SLL_ERROR_INVALID_MAGIC_NUMBER`
 - Implemented `SLL_NODE_TYPE_COMPLEX` and `SLL_ASSEMBLY_INSTRUCTION_TYPE_PUSH_COMPLEX` ([#262])
 - Implemented `SLL_OBJECT_TYPE_COMPLEX`, `SLL_INIT_COMPLEX`, `SLL_INIT_COMPLEX_STRUCT`, `sll_static_complex_zero`, `sll_complex_to_object` and `complex_type` ([#262])
 - Integer and floats can now be prefixed with the `+` sign for consistency